feat(payment): PAYPAL-1713 update after review
