+++ packages/core/src/app/payment/paymentMethod/getBraintreeAchValidationSchema.tsx
@@ -1,6 +1,6 @@
-import { Address, FormField, LanguageService } from '@bigcommerce/checkout-sdk';
+import { FormField, LanguageService } from '@bigcommerce/checkout-sdk';
 import { memoize } from '@bigcommerce/memoize';
-import { NumberSchema, object, string, StringSchema } from 'yup';
+import { object, string, StringSchema } from 'yup';
 
 export enum BraintreeAchFieldType {
     BusinessName = 'businessName',
@@ -28,24 +28,6 @@ export const BraintreeAchBankAccountValues = {
 
 export type BraintreeAchBankAccount = BraintreeAchFieldType | BraintreeAchAddressType;
 
-export const filterAddressToFormValues = (address?: Address) => {
-    const addressValues: Partial<Address> = {};
-
-    if (!address) return addressValues;
-
-    const values = Object.values(BraintreeAchAddressType);
-
-    values.forEach((key) => {
-        const value = address[key];
-
-        if (value !== undefined) {
-            addressValues[key] = value;
-        }
-    })
-
-    return addressValues;
-}
-
 export default memoize(function getBraintreeAchValidationSchema({
     formFieldData,
     language,
@@ -70,9 +52,21 @@ export default memoize(function getBraintreeAchValidationSchema({
         if (required) {
             if (requiredFieldErrorTranslationIds[id]) {
                 schema[id] = string().required(language.translate(`${requiredFieldErrorTranslationIds[id]}_required_error`));
+
+                if (id === BraintreeAchFieldType.AccountNumber) {
+                    schema[id] = schema[id]
+                        .matches(/^\d+$/, language.translate('payment.only_numbers_error', { label: language.translate('payment.account_number_label') }))
+                }
+
+                if (id === BraintreeAchFieldType.RoutingNumber) {
+                    schema[id] = schema[id]
+                        .matches(/^\d+$/, language.translate('payment.only_numbers_error', { label: language.translate('payment.account_routing_label') }))
+                        .min(8, language.translate('customer.min_error', { label: language.translate('payment.account_routing_label'), min: 8 }))
+                        .max(9, language.translate('customer.max_error', { label: language.translate('payment.account_routing_label'), max: 9 }))
+                }
             }
         }
 
         return schema;
-    }, {} as Record<string, StringSchema | NumberSchema>));
+    }, {} as { [key: string]: StringSchema }));
 })
